Bugfix: Tie the `save_post` action only to registered PTAN post types
The `save_post` action was being used to perform a publish action on
all post types, without regard to whether the post type was registered
with Publish to Apple News or not. Therefore, when a `save_post` action
triggered for a post type that was not registered to PTAN, the nonce
verification check failed, since the metabox was not being displayed
and therefore the nonce was not being generated. This commit modifies
the behavior of the `save_post` action to register separate `save_post`
action hooks for each supported post type.